Update the note when the moderation ends with success, by calling the GET on the REST API, and refresh the value in the DB. Also updates the UI accordingly.
